Ensure the macos compatibility that is claimed in the python wheel
The python wheel says 11.0 for arm64 and 10.7 otherwise, but the crate
has dependencies that build C code that could pull symbols that aren't
available in older version of macOS, because of building against a newer
SDK.

Although practically speaking, it was probably not happening.
Technically, though, rust code could be doing the same and that wouldn't
be detected.