Add wheels for arm64 macos (#1534)
* Use cp36 and abi3 tags on all platforms

All versions of python >= 3.2 are compatible with abi3. cp36 indicates
compat with >= 3.6.

The wheels module used to create wheels with py3_none on Windows, but
that was, as far as I know, a bug, and wheels with cp3*_abi3 work just
fine.

* Handle compiling glean_ffi.dll in setup.py

This avoids special-casing in the CI configuration, allowing to build
locally with less extra manual steps.

* Use rust compiler host triple instead of sys.platform

The rust host triple is its default target. Instead of relying on the
python platform, we now rely on that host triple to make decisions,
which is more likely to fit what we really want.

And now that we know the default target for the rust compiler, we can
also always explicitly pass it, such that the shared library path is
consistently under target/$target, which simplifies things.

At the same time, we stop assuming "linux" and "darwin" means x86_64.

* Allow to use an arbitry target when building

This allows a more general way to use a custom target, and use that for
Windows wheels on CI.

* Add a CI job building wheels for arm64 macos

* Ensure the macos compatibility that is claimed in the python wheel

The python wheel says 11.0 for arm64 and 10.7 otherwise, but the crate
has dependencies that build C code that could pull symbols that aren't
available in older version of macOS, because of building against a newer
SDK.

Although practically speaking, it was probably not happening.
Technically, though, rust code could be doing the same and that wouldn't
be detected.

* Use a py3-none tag on Windows after all

cp36-abi3 only works with pip >= 20, but we still have use-cases on pip
19.3.