Pick clang 11 and cctools from m-c build
clang and cctools selected from the following treeherder run:
https://treeherder.mozilla.org/jobs?repo=autoland&revision=11810f926917ae7e79b7fa156239b2cab7b0327c

This was picked from checking what artifacts another cross-compile build
on Treeherder picked.

This contains clang 11 and corresponding cctools.
This supports both x86_64 and aarch64 darwin. We only need the former
for now (but might use the latter eventually).

It's not clear how stable these URLs are long-term, but we'll work on
getting them into tooltool (or make sure we cache them in the Docker
container instead of installing at runtime)