RLB: API to pass in raw samples for timing distributions
Unfortunately this needs a new API instead of the already existing `accumulate_samples_signed`.
We want the RLB consumer (in this case: FOG) to not care about the unit,
so we expect nanosecond samples.
Internally we then do the right thing.

And because this is pure Rust we can rely on `u64` and just never need
to filter out negative samples.