Bug 1691073, Bug 1703095: Record errors from bindings
This is just a PoC PR to solicit feedback before fanning it out to all of the
metric types.

This just handles the string metric type for RLB and Python.

The Python bindings add a generic decorator that can be used to catch all
exceptions when recording a metric and converting them to a Glean error.