Implement an additional check to avoid crash due to faulty timezone offset
`Local::now` takes the user's timezone offset
and panics if it's not within a range of [-24, +24] hours.
This causes crashes in a small number of clients on Windows.

We can't determine the faulty clients
or the circumstancens under which this happens,
so the best we can do is have a workaround:

We try getting the time and timezone first,
then manually check that it is a valid timezone offset.
If it is, we proceed and use that time and offset.
If it isn't we fallback to UTC.

This has the small downside that it will use 2 calls to get the time,
but only on Windows.

See https://bugzilla.mozilla.org/show_bug.cgi?id=1611770.