Instrument timezone offset workaround
This requires more changes than I'd like,
but recording metrics internally requires a valid `Glean` to be passed around.
Because we use `local_now_with_offset` right when we construct a `Glean` instance,
we need to first finish that and _then_ report it.
In other places we have a `Glean` around and can pass it on.
So we split up the code and do the recording externally when used during
initialization, and internally whenever we can pass in a valid `Glean`.