Rust: Split glean-ffi up into 2 crates
glean-ffi carries all the no_mangle extern "C" definitions of our FFI
layer. This is getting included in mozilla-central soon.
However we can't have crates that are build as a `cdylib` as
dependencies in mozilla-central, as this will then produce said `cdylib`
during the build, which breaks at least on one build target.

We don't need that `cdylib` published, we only need it for builds we
ship ourselves: Kotlin, Swift & Python.

Therefore we split this part into a `glean-bundle` crate which itself
will generate a `cdylib` by the name `libglean_ffi`.
Our builds only need to reference that other workspace member, but don't
need to change what library they are loading.
Now `glean-ffi` is a plain ol' Rust crate and we can use that in
mozilla-central.