Rust: Force symbol re-export from libglean_ffi.
Workaround to force a re-export of the `no_mangle` symbols from `glean_ffi`

Due to how linking works and hides symbols the symbols from `glean_ffi` might not be
re-exported and thus not usable.
By forcing use of _at least one_ symbol in an exported function the functions will also be
rexported.
This is only required for debug builds (and `debug_assertions` is the closest thing we have to
check that).
In release builds we rely on LTO builds to take care of it.
As the tests will use libglean_ffi functionality running the tests
should ensure this actually happens.

See https://github.com/rust-lang/rust/issues/50007