Wait for the last upload thread to finish work before shutting down.
In tests, before we shut down the global Glean, we ensure that the last
upload thread finishes work.
That thread reads from the data path and also tries to store metrics
into the database. It needs Glean around for that to work.

In tests this should be a fast process.
At the end of individual tests one should not have more than a couple
pending pings around.