When resetting Glean in testing wait for the init task to finish
Waiting for initialization to finish means we know everything is set up
correctly.

In tests this is useful as we start and shutdown Glean frequently.
We might even be so fast as that init is not finished when we call
shutdown, but then we already try to start again.

This is not happening in production builds, as we only start Glean
exactly once per run and never reset it.
Tasks should be correctly queued and executed once init is done and the
queue is drained.

Note: This means we're not testing that behavior anymore in unit tests that use `test_reset_glean`.
Which IMO is fine;
we will need test coverage for the explicit delayed initialization of Glean in downstream users, e.g. FOG, on another layer of testing.