RLB: Reset global state after Glean is destroyed
If Glean fails to initialize the global state, containing a handle to
the uploader task, is not reset.
Subsequent resets expect this state to be valid and try to wait for any
remaining upload task. But if Glean never fully initialized there will
be none.
So we just reset that state after Glean is destroyed, as we know there
can't be any _real_ upload thread still hanging around.