Use the correct type for the `seq` field
This definition is only useful for documentation purposes, because [we are not using it internally](https://searchfox.org/glean/rev/42e5778f4aaf240330eae1ce5bebfa35e19642a0/glean-core/src/ping/mod.rs#71). We are not even using a field with the same name :-) We're using `sequence` internally, instead of `seq`.