Apply clippy suggestions everywhere
The new beta clippy found all of these, mostly this should be turning
`&self` into `self` when passing as an argument, when `self` is already `&Self`, so a reference.
That double-referencing will just go away anyway (and the compiler
should be smart enough to not make that costly at all).

I ran this to apply it all automatically:

     cargo +nightly clippy --verbose --all --all-targets --all-features -Z unstable-options --fix -- -D warnings