Issue #1717 - Stop setting MOZ_FETCHES_DIR
The run-task script will rm -rf $MOZ_FETCHES_DIR at the end, so since
this was being set to '/builds/worker/checkouts/src', we were actually
clobbering the entire checkouts dir each time.

Specifically, this included the 'docker-contexts' artifacts the decision
task generates, which breaks docker-image tasks.

It looks like MOZ_FETCHES_DIR is only used in commented out code, but
even still it shouldn't need to be set in the env of the Decision task
itself. Instead, it needs to be set in the env of the task(s) that are
going to be using it.