[Bug 1730473] Unit tests leaking pings to ingestion pipeline
This addresses the iOS tests that are leaking pings to the ingestion pipeline and are showing up in the Glean Debug View. This wraps any call to `resetGlean` with an OHTTPStub in order to intercept the request and prevent the network activity. Most test cases just reset Glean in the `setUp` function, and so the startup pings are now intercepted so they don't interfere with the tests or escape. In some cases, where we were already using stubs as we test contents of pings, the intent of the test was maintained and any setup calls to `resetGlean` were also wrapped in a way to not interfere with these tests' intents.