Implement `setUploadEnabled` using a callback mechanism
`setUploadEnabled` is a bit more complex than just flipping the `uploadEnabled` bit.

Before disabling upload we want any upload operation to be stopped.
The Metrics ping scheduler need to be cancelled too.

If upload is not actually changed no further action is required.

If upload is changed from on to off we need to trigger the uploader
once more to upload a deletion request ping.
If upload is changed from off to on we need to set the core metrics
again, which includes data from the application.

All of that should run in the dispatcher for 2 reasons:
* Not blocking the application
* Running in the expected order with other Glean recording operations

To keep most of current logic we can use callbacks on each of these cases.

The alternative to this would be:
* Store the required data for core metrics on the Rust side on
  initialization time.
* Store a general callback to trigger the uploader/cancel the uploader
  at initialization time.
* Make `setUploadEnabled` handle all of it in Rust using the previous
  set data/callbacks instead of relying on new ones.