Rename 'src -> vcs' in hardcoded paths in CI tasks
The previous commit updated Taskgraph to the latest version, which
contained a backwards incompatible change (introduced last summer) that
renamed the default checkout path from 'src' -> 'vcs'. This is causing
the build to fail.

Since this backwards incompatible change was introduced, we've since
started releasing to Pypi with proper version numbers and are now
following SemVer. So hopefully cases like this won't happen again.

(Ironically preventing cases like this is what the PR was trying to
solve).