Remove `rlb_flush_dispatcher` symbol
This was used to unblock the Rust dispatcher from foreign language bindings.
Now that we have _only_ the Rust dispatcher it's of course the one
that's going to be unblocked by `initialize`.
The functionality to automatically call this from `glean-ffi` was
removed when `glean-ffi` itself was removed.

This is thus just a cleanup of already unused functionality.