bug 1767037 - RLB: Add a new ctor to Event that takes a list of allowed event extra keys.
This enables runtime-defined extra keys for runtime-defined event metrics.

Alternatively to this we could give RLB a RuntimeExtraKeys like NoExtraKeys and
specialize the impl of EventMetric to not use ALLOWED_EXTRA or into_ffi_extra,
but given that runtime-defined metrics are a dev-only feature (at present)
maybe this is enough?