CI: Use host clang instead of Xcode-bundled clang
Using the Xcode-bundled one (which is the first one in the PATH when
compiling) everything fails when building for the host (build scripts,
tooling) with linker errors.
Instead of figuring out the right magic words, variables and paths the
easier way seems to just use /usr/bin/clang.

Note that it still requires the LIBRARY_PATH changes from the parent
commit.