RLB: Wait for uploader on shutdown using new `shutdown` mechanism
On shutdown we now wait up to 30s for the upload thread to finish.
This gets done just after we drain and shut down the dispatcher,
but before Glean is considered fully shutdown.

The callback MUST NOT put any new tasks on the dispatcher.
The callback SHOULD NOT block arbitrarily long.

The RLB implementation of that callback instructs a potential uploader
thread to end operation after it finished one upload.
It also blocks any new upload thread from getting started.
This means an in-flight upload can finish as usual, but no new one
is launched.
If the upload thread is currently sleeping, it gets one more upload.

This also adds a new test.
The only way to reliably test this is from the outside,
so the test invokes an example implementation and checks that the
`pending_pings` directory is empty after the program ends.
The uploader inside sleeps for a short period of time.
Without waiting for this uploader the main thread would end and kill the
uploader thread before it wakes up.
With waiting the uploader finishes correctly before the main thread
exits.