Instrument the time needed for a successful/failure upload of a ping
We want to measure this for every SDK,
but the actual upload is a multi-state process,
where the uploader first requests the upload and then later submits the
status of that upload.

In theory there can be multiple uploader threads in parallel,
though our own uploaders should run sequential only.
Due to this we need to keep track of in-flight uploads by document ID,
so we can later ensure we stop the correct timer.