RLB: Gracefully handle uploader thread start failure
There might be cases where we can't properly start the thread, e.g. when
out of memory (as it seems to happen on low-end 32bit versions of Firefox).
We now avoid the crash.
The data should be safely stored on disk and on the next upload it will
try again (and hopefully succeed).