iOS: Handle `didBecomeActiveNotification` to track first active state of an app
We used to always call `handleForegroundEvent` after initialize, when
setting up the lifecycle observer.
Howerver there are cases where the app starts up but is NOT made active.

The only way to really capture that is to listen to `didBecomeActiveNotification`.
Now `didBecomeActiveNotification` is also triggered after the user leaves an overlay
(popup, notification), but that's okay in our case:
`handleForegroundEvent` keeps track of the `active` status,
but the above cases don't trigger `didEnterBackgroundNotification`!