Add a configuration option to disable internal pings
This will mark the internal pings metrics, baseline and events as
disabled and will not send them.
Data recorded for these pings is still recorded, but never cleared out.

This setting should only be used for apps that definitely do not want
any of these pings and do not record any metrics of their own into those
pings.