Bug 1804711 - Part 2: Have submit_sync also send pings listed under schedules_pings.
While glean_parser will try to prevent the base-case of pings trying
to schedule _themselves_ when being sent, nothing in this stack takes
care of the problem of recursive cycles of pings.

So if we have PingA scheduled to be sent when PingB is sent, and
PingB is scheduled to be sent when PingC is sent, and PingC is
scheduled to be sent when PingA is sent... things get messy
indeed, and I think we might get stuck in a ping-sending loop.