Kotlin: Use experimental linter version to avoid `NewApi` false positive
For whatever reason a newer Kotlin triggers this error in the linter:

    glean-core/android/src/main/java/mozilla/telemetry/glean/Glean.kt:56: Error:
    Call requires API level 24 (current min is 21): java.lang.Iterable#forEach [NewApi]

which is not correct.
That code exists since forever in Glean and running on older Android
just fine.

Seems to be a known problem:
https://issuetracker.google.com/issues/185418482

So by using a later version of the linter we can avoid it.
Note: that version is not experimental!
It's a stable release, we just cannot upgrade to AGP 8.3.0 due to other
issues.