Collect samples of Rkv commit timing
Due to this being deep within the database we need to do things slightly
differently.

We will measure the time ourselves and keep the samples.
Then on _any_ ping assemble we record those values into the database
(essentially batching those timings).
Then it just goes the usual way.

We _do_ want the timings to persist, because we want to collect it
across a client's usage, which might include several application runs
but only a single metrics ping.

The downside is that recording the write times itself causes a write.