Persist ping-lifetime data after 1000 writes
We now automatically flush data to disk after 1000 writes into the
in-memory ping-lifetime map.

If `delay_ping_lifetime_io` is set to true, the data is now flushed to
disk in these cases:
* When `persist_ping_lifetime_data` is called by the application.
* At shutdown (that calls `persist_ping_lifetime_data`) automatically.
* Every 1000 writes
  * Calling `persist_ping_lifetime_data` resets the counter!

The downside with `delay_ping_lifetime_io` is the increased risk of
small windows of data loss:
Data is only persisted so often, so if the process crashes or is killed
in between the data is not persisted to disk and we will never know about.
We do our best to still persist frequently (after enough writes, on ping
submission, on background) and the application could decide to call
`persist_ping_lifetime_data` on their own as well, e.g. on idle.