Kotlin: Add a (blocked) task queue and unblock it after initializing
Note that while Glean is initializing (in glean-core on a separate thread),
we run all tasks from the early queue.
These then just add to the queue from Glean.

Now, in theory, this could slightly change the order in which certain
metric recordings are executed. I think. I can't come up with any sort
of proof or test or so on whether that's problematic.