Kotlin: Dispatch metric recording for event, object and timing distribution on the task queue
This allows us to delay executing those until Glean gets initialized and
thus the core library is loaded.
This is important for Fenix startup where we use `libxul` from Gecko,
but want it to be loaded by GeckoView code before Glean accesses it.

The three metric types for which this is currently implemented are the
three metric types that are currently recorded early during Fenix' startup (by Nimbus).

For the timing distribution this task dispatching is ONLY used for the
`accumulate` methods.
We can't do that for start/stop, because the timings are handled within glean-core.