bug 1928288 - Permit shutdowns to interrupt an RLB UploadManager Wait task
This permits must faster orderly shutdowns if we happen to be throttled.

This is acceptable because the point of having a 30s external timeout on upload
shutdown is to allow a session time to attempt upload of at least one ping.
If we are throttled, it's because we've sent more than one (default: 15) ping.