UniFFI: Pass library path when generating foreign-language bindings
We will be using macros soon, and that's when we need the library file
to parse out additional types.

Unfortunately currently maturin in particular determines library mode
based on whether there's _any_ UDL file in tree, which no way to
overwrite this (https://github.com/PyO3/maturin/issues/2617).
So we simply go look for the library and always pass it along.
We do that for every one of our targets, then we won't even need to pass
a new flag to uniffi-bindgen in the calling scripts.