(backported) MPS: Properly calculate rescheduled duration using local time and avoid re-scheduling anything immediately
This code changed with the recent chrono update and can cause rapid
rescheduling of metric pings due to wrong date and time math, that mixes
local, naive and UTC times.

We now do the time math _always_ in the local timezone, ensuring to aim
for a 4am submission of the metrics ping (in case it gets rescheduled).
We also avoid any 0-duration reschedule (immediate). If somehow our math
doesn't work out we schedule the metrics ping 24 hours into the future.