Kotlin: For lateinit metrics, check that the metric was instantiated before recording on it
This _should not_ happen, but apparently it does in Fenix. I have no
idea why or how to prevent it.
But we can mitigate it.
Downside: If the instance is not instantiated by the time `set` actually
runs we don't record any data. But then again before we just crashed,
which also didn't record data.