CI: Deploy pages using GhA
It only runs on pushes to the `main` branch.
With this we can drop the `gh-pages` branch which bloats the entire repository.

This does the minimal thing:

It builds Rust docs, Python docs & the books.
It then deploys them.

On GitHub Actions this does NOT do:

1. Spellcheck
2. Linkcheck
3. Swift docs

(1) and (2) are still done on CI as part of the usual CI from branches/PR runs.
We do not do them on merges to `main` though.

Swift docs (3) are a bit harder, as it requires Xcode and thus requires macOS.
That's a bit more expensive in terms of resources.
We could bring it back on GitHub Actions too if needed.
But it's not fully clear how much value those rendered docs bring.